Fix some nitpicks with datagen (#5319)
* Fix some nitpicks

* whoops
